finished refactoring using stores to share state between components
