+++ project/frontend/src/App.svelte
@@ -12,25 +12,10 @@
 		if (res.ok) {
 			todos = await res.json();
 			todoList.set(todos);
-			console.log(todos);
 		} else {
 			console.log("HTTP-Error: " + res.status);
 		}
 	});
-
-	async function updateTodoList() {
-		let newlist;
-		const response = await fetch(apiUrl, {
-			method: "GET",
-		});
-		if (response.ok) {
-			newlist = await response.json();
-		} else {
-			console.log("HTTP-Error: " + response.status);
-		}
-		console.log(newlist);
-		todoList.set(newlist);
-	}
 </script>
 
 <style>
@@ -58,9 +43,6 @@
 
 <main>
 	<h1>{title}</h1>
-	<TodoForm
-		txtPlaceholder="New ToDo task"
-		{apiUrl}
-		on:notify={updateTodoList} />
+	<TodoForm txtPlaceholder="New ToDo task" {apiUrl} />
 	<TodoList />
 </main>
